+++ tests/integrations/django/conftest.py
@@ -1,0 +1,33 @@
+import pytest
+import random
+
+from werkzeug.test import Client
+
+from tests.integrations.django.myapp.wsgi import application
+
+
+@pytest.fixture
+def client():
+    return Client(application)
+
+
+@pytest.fixture
+def use_django_caching(settings):
+    settings.CACHES = {
+        "default": {
+            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
+            "LOCATION": "unique-snowflake-%s" % random.randint(1, 1000000),
+        }
+    }
+
+
+@pytest.fixture
+def use_django_caching_with_middlewares(settings):
+    settings.CACHES = {
+        "default": {
+            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
+            "LOCATION": "unique-snowflake-%s" % random.randint(1, 1000000),
+        }
+    }
+    settings.MIDDLEWARE.insert(0, "django.middleware.cache.UpdateCacheMiddleware")
+    settings.MIDDLEWARE.append("django.middleware.cache.FetchFromCacheMiddleware")
