+++ tests/integrations/django/test_basic.py
@@ -2,10 +2,8 @@ from __future__ import absolute_import
 
 import json
 import pytest
-import pytest_django
 from functools import partial
 
-from werkzeug.test import Client
 from django import VERSION as DJANGO_VERSION
 from django.contrib.auth.models import User
 from django.core.management import execute_from_command_line
@@ -22,30 +20,7 @@ from sentry_sdk.consts import SPANDATA
 from sentry_sdk.integrations.django import DjangoIntegration
 from sentry_sdk.integrations.django.signals_handlers import _get_receiver_name
 from sentry_sdk.integrations.executing import ExecutingIntegration
-
-from tests.integrations.django.myapp.wsgi import application
-
-# Hack to prevent from experimental feature introduced in version `4.3.0` in `pytest-django` that
-# requires explicit database allow from failing the test
-pytest_mark_django_db_decorator = partial(pytest.mark.django_db)
-try:
-    pytest_version = tuple(map(int, pytest_django.__version__.split(".")))
-    if pytest_version > (4, 2, 0):
-        pytest_mark_django_db_decorator = partial(
-            pytest.mark.django_db, databases="__all__"
-        )
-except ValueError:
-    if "dev" in pytest_django.__version__:
-        pytest_mark_django_db_decorator = partial(
-            pytest.mark.django_db, databases="__all__"
-        )
-except AttributeError:
-    pass
-
-
-@pytest.fixture
-def client():
-    return Client(application)
+from tests.integrations.django.utils import pytest_mark_django_db_decorator
 
 
 def test_view_exceptions(sentry_init, client, capture_exceptions, capture_events):
