+++ tests/integrations/django/test_caching_spans.py
@@ -1,49 +1,18 @@
 import pytest
-import random
 
-from werkzeug.test import Client
+from sentry_sdk._compat import PY2
+from sentry_sdk.integrations.django import DjangoIntegration
+from tests.integrations.django.utils import pytest_mark_django_db_decorator
 
 try:
     from django.urls import reverse
 except ImportError:
     from django.core.urlresolvers import reverse
 
-from sentry_sdk.integrations.django import DjangoIntegration
-
-from tests.integrations.django.myapp.wsgi import application
-
-from sentry_sdk._compat import PY2
-
-
-@pytest.fixture
-def client():
-    return Client(application)
-
-
-@pytest.fixture
-def use_django_caching(settings):
-    settings.CACHES = {
-        "default": {
-            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
-            "LOCATION": "unique-snowflake-%s" % random.randint(1, 1000000),
-        }
-    }
-
-
-@pytest.fixture
-def use_django_caching_with_middlewares(settings):
-    settings.CACHES = {
-        "default": {
-            "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
-            "LOCATION": "unique-snowflake-%s" % random.randint(1, 1000000),
-        }
-    }
-    settings.MIDDLEWARE.insert(0, "django.middleware.cache.UpdateCacheMiddleware")
-    settings.MIDDLEWARE.append("django.middleware.cache.FetchFromCacheMiddleware")
-
 
 @pytest.mark.forked
-def test_cache_spans_disabled_middleware(
+@pytest_mark_django_db_decorator()
+def test_cache_spans_disabled_middleware_XXX(
     sentry_init,
     client,
     capture_events,
@@ -62,11 +31,8 @@ def test_cache_spans_disabled_middleware(
     )
     events = capture_events()
 
-    # Calling client.get() messes up the Django urlpatterns,
-    # so we reverse all urls, before we get them with client.get()
-    url = reverse("not_cached_view")
-    client.get(url)
-    client.get(url)
+    client.get(reverse("not_cached_view"))
+    client.get(reverse("not_cached_view"))
 
     (first_event, second_event) = events
     assert len(first_event["spans"]) == 0
@@ -74,6 +40,7 @@ def test_cache_spans_disabled_middleware(
 
 
 @pytest.mark.forked
+@pytest_mark_django_db_decorator()
 def test_cache_spans_disabled_decorator(
     sentry_init, client, capture_events, use_django_caching
 ):
@@ -89,11 +56,8 @@ def test_cache_spans_disabled_decorator(
     )
     events = capture_events()
 
-    # Calling client.get() messes up the Django urlpatterns,
-    # so we reverse all urls, before we get them with client.get()
-    url = reverse("cached_view")
-    client.get(url)
-    client.get(url)
+    client.get(reverse("cached_view"))
+    client.get(reverse("cached_view"))
 
     (first_event, second_event) = events
     assert len(first_event["spans"]) == 0
@@ -101,6 +65,7 @@ def test_cache_spans_disabled_decorator(
 
 
 @pytest.mark.forked
+@pytest_mark_django_db_decorator()
 def test_cache_spans_disabled_templatetag(
     sentry_init, client, capture_events, use_django_caching
 ):
@@ -116,11 +81,8 @@ def test_cache_spans_disabled_templatetag(
     )
     events = capture_events()
 
-    # Calling client.get() messes up the Django urlpatterns,
-    # so we reverse all urls, before we get them with client.get()
-    url = reverse("view_with_cached_template_fragment")
-    client.get(url)
-    client.get(url)
+    client.get(reverse("view_with_cached_template_fragment"))
+    client.get(reverse("view_with_cached_template_fragment"))
 
     (first_event, second_event) = events
     assert len(first_event["spans"]) == 0
@@ -128,6 +90,7 @@ def test_cache_spans_disabled_templatetag(
 
 
 @pytest.mark.forked
+@pytest_mark_django_db_decorator()
 def test_cache_spans_middleware(
     sentry_init,
     client,
@@ -149,11 +112,8 @@ def test_cache_spans_middleware(
     )
     events = capture_events()
 
-    # Calling client.get() messes up the Django urlpatterns,
-    # so we reverse all urls, before we get them with client.get()
-    url = reverse("not_cached_view")
-    client.get(url)
-    client.get(url)
+    client.get(reverse("not_cached_view"))
+    client.get(reverse("not_cached_view"))
 
     (first_event, second_event) = events
     assert len(first_event["spans"]) == 1
@@ -181,6 +141,7 @@ def test_cache_spans_middleware(
 
 
 @pytest.mark.forked
+@pytest_mark_django_db_decorator()
 def test_cache_spans_decorator(sentry_init, client, capture_events, use_django_caching):
     sentry_init(
         integrations=[
@@ -194,11 +155,8 @@ def test_cache_spans_decorator(sentry_init, client, capture_events, use_django_c
     )
     events = capture_events()
 
-    # Calling client.get() messes up the Django urlpatterns,
-    # so we reverse all urls, before we get them with client.get()
-    url = reverse("cached_view")
-    client.get(url)
-    client.get(url)
+    client.get(reverse("cached_view"))
+    client.get(reverse("cached_view"))
 
     (first_event, second_event) = events
     assert len(first_event["spans"]) == 1
@@ -226,6 +184,7 @@ def test_cache_spans_decorator(sentry_init, client, capture_events, use_django_c
 
 
 @pytest.mark.forked
+@pytest_mark_django_db_decorator()
 def test_cache_spans_templatetag(
     sentry_init, client, capture_events, use_django_caching
 ):
@@ -241,11 +200,8 @@ def test_cache_spans_templatetag(
     )
     events = capture_events()
 
-    # Calling client.get() messes up the Django urlpatterns,
-    # so we reverse all urls, before we get them with client.get()
-    url = reverse("view_with_cached_template_fragment")
-    client.get(url)
-    client.get(url)
+    client.get(reverse("view_with_cached_template_fragment"))
+    client.get(reverse("view_with_cached_template_fragment"))
 
     (first_event, second_event) = events
     assert len(first_event["spans"]) == 1
